Address review feedback: extract ProtobufLanguage and add missing keywords
