+++ ui/ui-app/src/editors/registerLanguages.ts
@@ -1,5 +1,5 @@
 import { Monaco } from "@monaco-editor/react";
-import { registerProto } from "./ProtoEditor.tsx";
+import { registerProto } from "./ProtobufLanguage.ts";
 import { registerGraphQL } from "./GraphQLLanguage.ts";
 
 export const registerCustomLanguages = (monaco: Monaco) => {
